refactor: reorganize project structure and improve person model
- Move exception handling classes to dedicated exception package
- Move person model classes to domain/model/person package
- Update person provider and related generators
- Add new Id model class
- Fix typo in inbound package name
