--- random-data-api/src/main/java/com/random/data/adapter/outbound/provider/person/model/Id.java
@@ -1,6 +1,0 @@
-package com.random.data.adapter.outbound.provider.person.model;
-
-public record Id(
-        String name, String value
-)
-{ }
