+++ random-data-api/src/main/java/com/random/data/domain/model/person/Id.java
@@ -1,0 +1,6 @@
+package com.random.data.domain.model.person;
+
+public record Id(
+        String name, String value
+)
+{ }
